Update README.md
This GitHub repository contains the code and resources for a machine learning project that involves the classification and regression analysis of COVID-19 data. The project was conducted as part of an internship at Prinston Smart Engineers, where machine learning algorithms were applied to a COVID-19 dataset sourced from Kaggle. The project was executed using Google Colab for ease of collaboration and cloud-based computation.